backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-generativeai>=0.8.0
sqlalchemy==2.0.23
apscheduler==3.10.4
python-dotenv==1.0.0
websockets==12.0
pydantic==2.5.0

